Set IP address when interface is up, but delays routes
Some applications need to bind to an IP address. Sometimes, these applications
are started before the cable is connected. Support this behavior by setting IP
for interface with static configuration when it is seen as UP, and routes etc.
when cable is plugged in.

Added the new state GOT_IP_STATIC_UP. It will be used to only set the IP of
interfaces where the iface is up, but L1 down

Set the state accoriding to OPERSTATE on start-up

States are set correctly

When interface is seen as IFF_UP (i.e., cable not connected), then only set
address

Have a seprate state for interfaces that are IFF_UP, after address has been
changed

Only set IP adress when interface is detected as UP during start-up

Restored some state. Put IFF_RUNNING handling in a seprate clause

IFF_UP for interfaces that come UP are now working. Needs to detect cable
disconnect now

Interfaces can now move from UP to RUNNING

Interfaces are properly configured when switching from UP to RUNNING during
operation

Dont touch the IP address of interfaces with statc IPs (when they go down)

Clean up when link is unplugged

PPP are set to RUNNING, tb[OPERSTATE] is not correct