[FIX] Configure multihomed routing properly
The way I configured the multihomed routing was due to my lack of experience
with the topic. Instead of keeping duplicate routes etc., it is much better to
keep all routes related to one interface/address in one table, and then use
policy routing for routing lookups. This commit takes care of that.

The gateway and network route is added to the interface-specific table. We then
create three rules pointing to this table - one that matches source address, one
where destionation is network and one for trafic not bound to a socket.