Delete dirty rules
There are a number of reasons for why netlink messages might get lost, it is
after all an unreliable protocol. When a link is deleted, address and routes are
deleted by the kernel. However, we are responsible for deleting the rule. If
this for some reason should fail (or DELLINK is lost), we might be left with a
rule pointing to an incorrect routing table. This is not good. So whenever an
interface comes up, check for any rules matching the address (either source or
destination) and delete that rule.

Next step is to add support for multiple interfaces with the same address.