[FIX] Sanitise default traffic rules
Check and delete incorrect default traffic rules. A rule is defined as incorrect
when:

* There exists more rules with the same priority. Priority is now 91000 + metric
and table is always metric. Thus, pref 91000 + X will always point to table X.
Having multiple of this rule is clearly incorrect. Detect this by storing the
pref of the previous rule. If the pref of the current rule matches, delete.

* If the first rule does not match, check that there exists an interface with
that metric. If not, delete the rule.